+++ NEWS.md
@@ -5,3 +5,4 @@
 - add non-breaking space to `make_gloss_list()` output
 - add compatibility to the `latex` outputs
 - fill non-specified by user glosses with values from `lingglosses::glosses`
+- add color annotation for glosses without definition or duplicated glosses with the argument `annotate_definitionless`